Add all tests to one file
